requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP clients
# Constrain httpx to a version compatible with ollama (requires <0.26.0)
httpx>=0.25.2,<0.26.0
requests==2.31.0

# AI/ML - Ollama
ollama==0.1.6

# Tools
duckduckgo-search==4.1.1

# Google Calendar (optional - only if you use it)
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Retry logic
tenacity==8.2.3

# Environment variables
python-dotenv==1.0.0
import dateparser